+++ R/pseudo_heatmap.R
@@ -68,8 +68,7 @@ pseudo_heatmap <- function(
   fate_names <- colnames(pseudo)
   
   ## extract gene expression data
-  gene_expression <- t(assay(sce, assay_name))
-  gene_expression <- as.data.frame(gene_expression[, features])
+  gene_expression <- as.data.frame(t(assay(sce, assay_name)[features, , drop = FALSE]))
   colnames(gene_expression) <- features
   
   ## prepare lineage input